+++ index.html
@@ -45,7 +45,7 @@
 
 		<input type="button" id="open" onclick="open3D()" value="Open Chat" hidden></input>
 
-	  <script src="/socket.io/socket.io.js"></script>
+	  <script src="/node_modules/socket.io-client/dist/socket.io.js"></script>
 	  <script src="https://webrtc.github.io/adapter/adapter-latest.js"></script>
 	  <script src="client.js"></script>
 		<script src="three.js"></script>
