+++ package.json
@@ -6,6 +6,6 @@
   "license": "SEE LICENSE IN <filename>",
   "dependencies": {
     "node-static": "^0.7.10",
-    "socket.io": "^1.2.0"
+    "socket.io": "^1.7.4"
   }
 }
